refactor: mirror the library module structure in the example app
