+++ apps/example/src/routers/advanced/inheritance.ts
