+++ apps/example/src/routers/decorators/validators/comparison.ts
@@ -1,0 +1,88 @@
+import express, { Router } from 'express'
+import { bindingCargo, getCargo, Body, Equal, NotEqual, IsTrue, IsFalse, OneOf, Validate } from 'express-cargo'
+
+const router: Router = express.Router()
+
+class EqualExample {
+    @Body()
+    @Equal(3)
+    number!: number
+
+    @Body()
+    @Equal('text')
+    string!: string
+
+    @Body()
+    @Equal(true)
+    boolean!: boolean
+}
+
+router.post('/equal', bindingCargo(EqualExample), (req, res) => {
+    const cargo = getCargo<EqualExample>(req)
+    res.json(cargo)
+})
+
+class NotEqualExample {
+    @Body()
+    @NotEqual(3)
+    number!: number
+
+    @Body()
+    @NotEqual('text')
+    string!: string
+
+    @Body()
+    @NotEqual(true)
+    boolean!: boolean
+}
+
+router.post('/not-equal', bindingCargo(NotEqualExample), (req, res) => {
+    const cargo = getCargo<NotEqualExample>(req)
+    res.json(cargo)
+})
+
+class IsTrueExample {
+    @Body()
+    @IsTrue()
+    booleanValue!: boolean
+}
+
+router.post('/is-true', bindingCargo(IsTrueExample), (req, res) => {
+    const cargo = getCargo<IsTrueExample>(req)
+    res.json(cargo)
+})
+
+class IsFalseExample {
+    @Body()
+    @IsFalse()
+    booleanValue!: boolean
+}
+
+router.post('/is-false', bindingCargo(IsFalseExample), (req, res) => {
+    const cargo = getCargo<IsFalseExample>(req)
+    res.json(cargo)
+})
+
+class OneOfExample {
+    @Body()
+    @OneOf(['js', 'ts', 'html', 'css'])
+    language!: string
+}
+
+router.post('/one-of', bindingCargo(OneOfExample), (req, res) => {
+    const cargo = getCargo<OneOfExample>(req)
+    res.json(cargo)
+})
+
+class ValidateExample {
+    @Body()
+    @Validate(email => (email as string).split('@').length === 2)
+    email!: string
+}
+
+router.post('/validate', bindingCargo(ValidateExample), (req, res) => {
+    const cargo = getCargo<ValidateExample>(req)
+    res.json(cargo)
+})
+
+export default router
